requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==3.2.12
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jsonapi==5.0.0
djangorestframework-xml==2.0.0
drf-json-api==0.1.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
pytz==2020.1
requests==2.27.1
simplejson==3.17.6
sqlparse==0.3.1
uritemplate==4.1.1
urllib3==1.26.9
pre-commit==2.18.1
